Move commitment test
