--- suite/spdz2k/src/test/java/dk/alexandra/fresco/suite/spdz2k/protocols/computations/TestSpdz2kCommitmentComputation.java
@@ -1,95 +1,0 @@
-package dk.alexandra.fresco.suite.spdz2k.protocols.computations;
-
-import static org.junit.Assert.assertArrayEquals;
-import static org.junit.Assert.assertEquals;
-
-import dk.alexandra.fresco.framework.Application;
-import dk.alexandra.fresco.framework.TestThreadRunner.TestThread;
-import dk.alexandra.fresco.framework.TestThreadRunner.TestThreadFactory;
-import dk.alexandra.fresco.framework.builder.numeric.ProtocolBuilderNumeric;
-import dk.alexandra.fresco.framework.network.Network;
-import dk.alexandra.fresco.framework.sce.evaluator.EvaluationStrategy;
-import dk.alexandra.fresco.framework.util.AesCtrDrbg;
-import dk.alexandra.fresco.lib.generic.CommitmentComputation;
-import dk.alexandra.fresco.suite.ProtocolSuiteNumeric;
-import dk.alexandra.fresco.suite.spdz2k.AbstractSpdz2kTest;
-import dk.alexandra.fresco.suite.spdz2k.Spdz2kProtocolSuiteK64;
-import dk.alexandra.fresco.suite.spdz2k.datatypes.CompUInt128;
-import dk.alexandra.fresco.suite.spdz2k.datatypes.CompUInt128Factory;
-import dk.alexandra.fresco.suite.spdz2k.datatypes.CompUIntFactory;
-import dk.alexandra.fresco.suite.spdz2k.resource.Spdz2kResourcePool;
-import dk.alexandra.fresco.suite.spdz2k.resource.Spdz2kResourcePoolImpl;
-import dk.alexandra.fresco.suite.spdz2k.resource.storage.Spdz2kDummyDataSupplier;
-import dk.alexandra.fresco.suite.spdz2k.resource.storage.Spdz2kOpenedValueStoreImpl;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Random;
-import java.util.function.Supplier;
-import org.junit.Test;
-
-public class TestSpdz2kCommitmentComputation extends
-    AbstractSpdz2kTest<Spdz2kResourcePool<CompUInt128>> {
-
-  @Test
-  public void testCommitmentTwo() {
-    runTest(new TestCommitment<>(), EvaluationStrategy.SEQUENTIAL_BATCHED, 2);
-  }
-
-  @Test
-  public void testCommitmentThree() {
-    runTest(new TestCommitment<>(), EvaluationStrategy.SEQUENTIAL_BATCHED, 3);
-  }
-
-  @Override
-  protected Spdz2kResourcePool<CompUInt128> createResourcePool(int playerId, int noOfParties,
-      Supplier<Network> networkSupplier) {
-    CompUIntFactory<CompUInt128> factory = new CompUInt128Factory();
-    Spdz2kResourcePool<CompUInt128> resourcePool =
-        new Spdz2kResourcePoolImpl<>(
-            playerId,
-            noOfParties, null,
-            new Spdz2kOpenedValueStoreImpl<>(),
-            new Spdz2kDummyDataSupplier<>(playerId, noOfParties, factory.createRandom(), factory),
-            factory);
-    resourcePool.initializeJointRandomness(networkSupplier, AesCtrDrbg::new, 32);
-    return resourcePool;
-  }
-
-  @Override
-  protected ProtocolSuiteNumeric<Spdz2kResourcePool<CompUInt128>> createProtocolSuite() {
-    return new Spdz2kProtocolSuiteK64();
-  }
-
-  private static class TestCommitment<ResourcePoolT extends Spdz2kResourcePool<CompUInt128>>
-      extends TestThreadFactory<ResourcePoolT, ProtocolBuilderNumeric> {
-
-    @Override
-    public TestThread<ResourcePoolT, ProtocolBuilderNumeric> next() {
-      return new TestThread<ResourcePoolT, ProtocolBuilderNumeric>() {
-        @Override
-        public void test() {
-          int noParties = conf.getResourcePool().getNoOfParties();
-          List<byte[]> inputs = new ArrayList<>();
-          Random random = new Random(42);
-          for (int i = 1; i <= noParties; i++) {
-            byte[] bytes = new byte[32];
-            random.nextBytes(bytes);
-            inputs.add(bytes);
-          }
-          Application<List<byte[]>, ProtocolBuilderNumeric> testApplication =
-              root -> new CommitmentComputation(
-                  conf.getResourcePool().getCommitmentSerializer(),
-                  inputs.get(root.getBasicNumericContext().getMyId() - 1), noParties,
-                  conf.getResourcePool().getLocalRandomGenerator())
-                  .buildComputation(root);
-          List<byte[]> actual = runApplication(testApplication);
-          assertEquals(inputs.size(), actual.size());
-          for (int i = 0; i < actual.size(); i++) {
-            assertArrayEquals(inputs.get(i), actual.get(i));
-          }
-        }
-      };
-    }
-  }
-
-}
